refactor prefill router without moving guardrails
